+++ app/config.ts
@@ -138,6 +138,11 @@ export const ROUTES: RouteInfo[] = [
     description: `Example: sorting data with ${PRODUCT_NAME}`,
   },
   {
+    href: '/examples/column-dragging-and-toggling',
+    title: 'Column dragging and toggling',
+    description: `Example: dragging & toggling ${PRODUCT_NAME} columns`,
+  },
+  {
     href: '/examples/infinite-scrolling',
     title: 'Infinite scrolling',
     description: `Example: how to implement infinite scrolling with ${PRODUCT_NAME}`,