+++ package/index.ts
@@ -1,5 +1,6 @@
 'use client';
 
 export { DataTable } from './DataTable';
+export * from './hooks/useDragToggleColumns';
 export * from './types';
 export * from './utils';