+++ package/types/DataTableProps.ts
@@ -83,6 +83,12 @@ export type DataTableProps<T = Record<string, unknown>> = {
   defaultColumnProps?: DataTableDefaultColumnProps<T>;
 
   /**
+   * If you want to use drag and drop as well as toggle to reorder and toggle columns
+   * provide a unique key which will be used to store the column order in localStorage.
+   */
+  storeColumnsKey?: string | undefined;
+
+  /**
    * A default render function for all columns.
    * Accepts the current record, its index in `records` and the column `accessor` as
    * arguments and returns a React node (remember that a string is a valid React node too).